Planilla/requirements.txt
flet
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
googleapis-common-protos
oauth2client<4.0.0
setuptools
wheel
qrcode
pypng